MIG: migracion de codigo a V11, soporte solo para archivos .pk12, quitar funciones de X509 REF: limpieza de codigo, seguir guideslines de odoo REM: eliminar archivos imnecesarios
