+++ AUTHORS.txt
@@ -1,0 +1,2 @@
+2015 BMyA SA - Blanco Martín & Asociados
+2017-TODAY La Otra Opcion
