+++ __manifest__.py
@@ -1,29 +1,11 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2015  BMyA SA  (http://blancomartin.cl)
-#    All Rights Reserved.
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+
 {
     "name": """User Signature Key""",
-    'version': '9.0.0.0.1',
+    'version': '1.0.0.0',
     'category': 'Utilities',
     'sequence': 12,
-    'author':  'BMyA SA - Blanco Martín & Asociados',
+    'author':  'La Otra Opcion, BMyA SA - Blanco Martín & Asociados',
     'website': 'http://blancomartin.cl',
     'license': 'AGPL-3',
     'summary': '',
@@ -34,9 +16,7 @@ sign electronic documents.
 """,
     'external_dependencies': {
         'python': [
-            'M2Crypto',
             'base64',
-            'cStringIO',
             'OpenSSL'
         ]
     },
@@ -44,9 +24,8 @@ sign electronic documents.
         'base',
     ],
     'data': [
-        'views/user_signature_tab.xml',
-        'views/company_signature_tab.xml',
-        # 'security/ir.model.access.csv'
+        'views/res_users_view.xml',
+        'views/res_company_view.xml',
     ],
     'installable': True,
     'auto_install': False,
