--- security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_user_signature_key,access_user_signature_key,model_user_signature_key,,1,0,0,0
